+++ djangocms_alias/forms.py
@@ -4,10 +4,12 @@ from django.contrib.admin.widgets import (
     AdminTextInputWidget,
     RelatedFieldWidgetWrapper,
 )
+from django.contrib.sites.models import Site
 from django.db import transaction
 from django.utils.translation import gettext_lazy as _
 
 from cms.models import CMSPlugin, Placeholder
+from cms.utils import get_current_site
 from cms.utils.permissions import (
     get_model_permission_codename,
     has_plugin_permission,
@@ -261,6 +263,10 @@ class Select2Mixin:
         )
 
 
+class SiteSelectWidget(Select2Mixin, forms.Select):
+    pass
+
+
 class CategorySelectWidget(Select2Mixin, forms.Select):
     pass
 
@@ -277,6 +283,13 @@ class AliasSelectWidget(Select2Mixin, forms.TextInput):
 
 
 class AliasPluginForm(forms.ModelForm):
+    site = forms.ModelChoiceField(
+        label=_("Site"),
+        queryset=Site.objects.all(),
+        widget=SiteSelectWidget(attrs={"data-placeholder": _("Select site")}),
+        required=False,
+    )
+
     category = forms.ModelChoiceField(
         label=_('Category'),
         queryset=Category.objects.all(),
@@ -302,10 +315,12 @@ class AliasPluginForm(forms.ModelForm):
         super().__init__(*args, **kwargs)
         if self.instance and self.instance.pk:
             self.fields['category'].initial = self.instance.alias.category_id
+        self.fields['site'].initial = get_current_site()
 
     class Meta:
         model = AliasPlugin
         fields = (
+            'site',
             'category',
             'alias',
             'template',